just organize code
